+++ docente/viaggioDiariaFattaSave.php
@@ -17,6 +17,7 @@ if(isset($_POST)) {
 	$descrizione = escapePost('descrizione');
 	$giorni_senza_pernottamento = $_POST['giorni_senza_pernottamento'];
 	$giorni_con_pernottamento = $_POST['giorni_con_pernottamento'];
+	$ore = $_POST['ore'];
 	$commento = escapePost('commento');
 
 	// il dirigente gestisce anche i commenti
@@ -32,10 +33,11 @@ if(isset($_POST)) {
 				if ($attuale != null) {
 					$giorni_senza_pernottamento_originali = $attuale['giorni_senza_pernottamento'];
 					$giorni_con_pernottamento_originali = $attuale['giorni_con_pernottamento'];
+					$ore_originali = $attuale['ore'];
 				}
 
-				dbExec("INSERT INTO viaggio_diaria_fatta_commento (commento, giorni_senza_pernottamento_originali, giorni_con_pernottamento_originali, viaggio_diaria_fatta_id) VALUES('$commento', '$giorni_senza_pernottamento_originali', '$giorni_con_pernottamento_originali', $id);");
-				info("inserito viaggio_diaria_fatta_commento commento=$commento giorni_senza_pernottamento_originali=$giorni_senza_pernottamento_originali giorni_con_pernottamento_originali=$giorni_con_pernottamento_originali viaggio_diaria_fatta_id=$id");
+				dbExec("INSERT INTO viaggio_diaria_fatta_commento (commento, giorni_senza_pernottamento_originali, giorni_con_pernottamento_originali, ore_originali, viaggio_diaria_fatta_id) VALUES('$commento', '$giorni_senza_pernottamento_originali', '$giorni_con_pernottamento_originali', '$ore_originali', $id);");
+				info("inserito viaggio_diaria_fatta_commento commento=$commento giorni_senza_pernottamento_originali=$giorni_senza_pernottamento_originali giorni_con_pernottamento_originali=$giorni_con_pernottamento_originali ore_originali=$ore_originali viaggio_diaria_fatta_id=$id");
 			} else {
 				// aggiorna il commento ma lascia i giorni originali inalterati
 				dbExec("UPDATE viaggio_diaria_fatta_commento SET commento = '$commento' WHERE viaggio_diaria_fatta_id=$id;");
@@ -44,7 +46,7 @@ if(isset($_POST)) {
 		} else {
 			// se e' vuoto decide se aggiornarlo o cancellarlo
 			if ($commento_id != null) {
-				dbExec("UPDATE viaggio_diaria_fatta_commento SET commento = '$commento', giorni_senza_pernottamento_originali = null, giorni_con_pernottamento_originali = null WHERE viaggio_diaria_fatta_id=$id;");
+				dbExec("UPDATE viaggio_diaria_fatta_commento SET commento = '$commento', giorni_senza_pernottamento_originali = null, giorni_con_pernottamento_originali = null, ore_originali = NULL WHERE viaggio_diaria_fatta_id=$id;");
 				info("aggiornato vuoto viaggio_diaria_fatta_commento commento=$commento viaggio_diaria_fatta_id=$id");
 			}
 		}
@@ -52,17 +54,16 @@ if(isset($_POST)) {
 
 	// in ogni caso aggiorna o inserisce il viaggio
 	if ($id > 0) {
-		dbExec("UPDATE viaggio_diaria_fatta SET data_partenza = '$data_partenza', descrizione = '$descrizione', giorni_senza_pernottamento = '$giorni_senza_pernottamento', giorni_con_pernottamento	 = '$giorni_con_pernottamento' WHERE id = '$id';");
+		dbExec("UPDATE viaggio_diaria_fatta SET data_partenza = '$data_partenza', descrizione = '$descrizione', giorni_senza_pernottamento = '$giorni_senza_pernottamento', giorni_con_pernottamento = '$giorni_con_pernottamento', ore = '$ore' WHERE id = '$id';");
 		info("aggiornata viaggio_diaria_fatta id=$id data_partenza=$data_partenza descrizione=$descrizione giorni_senza_pernottamento=$giorni_senza_pernottamento giorni_con_pernottamento=$giorni_con_pernottamento docente_id=$docente_id");
 	} else {
-		dbExec("INSERT INTO viaggio_diaria_fatta (data_partenza, descrizione, giorni_senza_pernottamento, giorni_con_pernottamento, docente_id, anno_scolastico_id) VALUES('$data_partenza', '$descrizione', '$giorni_senza_pernottamento', '$giorni_con_pernottamento', '$docente_id', '$__anno_scolastico_corrente_id');");
+		dbExec("INSERT INTO viaggio_diaria_fatta (data_partenza, descrizione, giorni_senza_pernottamento, giorni_con_pernottamento, ore, docente_id, anno_scolastico_id) VALUES('$data_partenza', '$descrizione', '$giorni_senza_pernottamento', '$giorni_con_pernottamento', '$ore', '$docente_id', '$__anno_scolastico_corrente_id');");
 		$ore_previste_attivita_id = dblastId();
-		info("aggiunto viaggio_diaria_fatta id=$id descrizione=$descrizione data_partenza=$data_partenza giorni_senza_pernottamento=$giorni_senza_pernottamento giorni_con_pernottamento=$giorni_con_pernottamento docente_id=$docente_id");
+		info("aggiunto viaggio_diaria_fatta id=$id descrizione=$descrizione data_partenza=$data_partenza giorni_senza_pernottamento=$giorni_senza_pernottamento giorni_con_pernottamento=$giorni_con_pernottamento ore='$ore' docente_id=$docente_id");
 	}
 
-	// La diaria non influenza il conteggio delle ore
-	// require_once '../docente/oreDovuteAggiornaDocente.php';
-	// oreFatteAggiornaDocente($__docente_id);
+	require_once '../docente/oreDovuteAggiornaDocente.php';
+	oreFatteAggiornaDocente($__docente_id);
 }
 
 ?>